# model/src/main/java/dk/dbc/promat/service/dto/ServiceErrorCode.java
/*
 * Copyright Dansk Bibliotekscenter a/s. Licensed under GPLv3
 * See license text in LICENSE.txt or at https://opensource.dbc.dk/licenses/gpl-3.0/
 */

package dk.dbc.promat.service.dto;

public enum ServiceErrorCode {
    EXTERNAL_SERVICE_ERROR,
    INVALID_REQUEST,
    FAUST_IN_USE,
    INVALID_STATE,
    FAILED,
    NOT_FOUND,
    FORBIDDEN
}