# service/docs/openapi.yaml
openapi: 3.0.1
info:
  title: Promat service
  description: >-
    Provides backend services for the Promat system.
  license:
    name: GPL-3.0
    url: 'https://github.com/DBCDK/promat-service/blob/master/LICENCE.txt'
  version: 1.0-SNAPSHOT
...
paths:
  /reviewers:
    get:
      tags:
        - reviewers
      summary: Returns list of reviewers
      operationId: getReviewers
      parameters:
        - name: deadline
          in: query
          required: false
          description: Optional deadline date as defined by RFC3339,
                       which will return reviewers with their workloads
                       for the week of the deadline as well as for the
                       weeks before and after the deadline.
          schema:
            type: string
            format: date
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReviewerList'
        '400':
          description: bad request, invalid deadline date
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
    post:
      tags:
        - reviewers
      summary: Create reviewer
      operationId: createReviewer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewerRequest'
      responses:
        '201':
          description: reviewer successfully created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviewer'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
        '500':
          description: On failure from external CULR service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
  /reviewers/{id}:
    get:
      tags:
        - reviewers
      summary: Returns a reviewer with a given ID
      operationId: getReviewer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            description: Reviewer ID
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviewer'
        '404':
          description: Reviewer not found
    put:
      tags:
        - reviewers
      summary: Update reviewer
      operationId: updateReviewer
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            description: Reviewer ID
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReviewerRequest'
      responses:
        '200':
          description: reviewer successfully updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Reviewer'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
        '404':
          description: Reviewer not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
        '500':
          description: On failure from external CULR service
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ServiceError'
...
components:
  schemas:
    Address:
      type: object
      properties:
        address1:
          type: string
        address2:
          type: string
        zip:
          type: string
        city:
          type: string
      required:
        - address1
        - zip
        - city
    ReviewerList:
      type: object
      properties:
        numFound:
          type: integer
        reviewers:
          type: array
          items:
            $ref: '#/components/schemas/Reviewer'
      required:
        - numFound
        - reviewers
    Reviewer:
      type: object
      properties:
        id:
          type: integer
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        institution:
          type: string
        paycode:
          type: integer
        hiatus_begin:
          type: string
          format: date
        hiatus_end:
          type: string
          format: date
        subjects:
          type: array
          items:
            $ref: '#/components/schemas/Subject'
        accepts:
          type: array
          items:
            $ref: '#/components/schemas/ReviewerAccepts'
        weekWorkload:
          type: integer
          description: workload for the week of a deadline
        weekBeforeWorkload:
          type: integer
          description: workload for the week before a deadline
        weekAfterWorkload:
          type: integer
          description: workload for the week after a deadline
        note:
          type: string
          description: editors notes on the reviewer
        capacity:
          type: integer
          description: maximum requested workload per week
      required:
        - id
        - firstName
        - lastName
        - email
        - address
        - institution
        - paycode
        - subjects
        - accepts
    ReviewerRequest:
      type: object
      properties:
        cprNumber:
          type: string
        firstName:
          type: string
        lastName:
          type: string
        email:
          type: string
        phone:
          type: string
        address:
          $ref: '#/components/schemas/Address'
        institution:
          type: string
        paycode:
          type: integer
        hiatus_begin:
          type: string
          format: date
        hiatus_end:
          type: string
          format: date
        subjects:
          type: array
          items:
            type: integer
          minItems: 0
          nullable: true
          description: IDs of subjects covered by this reviewer
        accepts:
          type: array
          items:
            $ref: '#/components/schemas/ReviewerAccepts'
        note:
          type: string
        capacity:
          type: integer
        active:
          type: boolean
      required:
        - cprNumber
        - firstName
        - lastName
        - email
        - paycode
...
    ServiceError:
      type: object
      properties:
        code:
          $ref: '#/components/schemas/ServiceErrorCode'
        cause:
          type: string
          description: Specific cause of the error,
                       for example the name of an invalid query parameter.
        details:
          type: string
      required:
        - code
        - cause
    ServiceErrorCode:
      type: string
      enum:
        - EXTERNAL_SERVICE_ERROR  # Error when communicating with external service
        - INVALID_REQUEST       # Request body contains invalid fields or id's
        - FAUST_IN_USE          # Active case for this/these faust/fausts exists
        - INVALID_STATE         # Change to a new state was not possible or invalid
        - FAILED                # Some error occurred
        - NOT_FOUND             # Case, Task or other item was not found
        - FORBIDDEN             # Operation was forbidden or impossible
...
    Subject:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        parentId:
          type: integer
      required:
        - id
        - name
...
    ReviewerAccepts:
      type: string
      enum:
        - 'BKM'
        - 'BOOK'
        - 'EBOOK'
        - 'EXPRESS'
        - 'MOVIE'
        - 'MULTIMEDIA'
        - 'PS4'
        - 'PS5'
        - 'XBOX_ONE'
        - 'XBOX_SERIES_X'
        - 'NINTENDO_SWITCH'
        - 'BIWEEKLY_ONLY'
...